requirements.txt
apt-clone==0.2.1
apturl==0.5.2
backcall==0.1.0
beautifulsoup4==4.8.1
bleach==3.1.0
Brlapi==0.6.6
bs4==0.0.1
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
colorama==0.4.1
command-not-found==0.3
configobj==5.0.6
cssselect2==0.2.2
cupshelpers==1.0
decorator==4.4.1
defer==1.0.6
defusedxml==0.6.0
docutils==0.15.2
EbookLib==0.17.1
feedparser==5.2.1
html5lib==1.0.1
httplib2==0.9.2
idna==2.8
ipython==7.10.0
ipython-genutils==0.2.0
jedi==0.15.1
louis==3.5.0
lxml==4.4.2
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
nemo-emblems==4.2.0
netifaces==0.10.4
onboard==1.4.1
PAM==0.4.2
parso==0.5.1
pbr==5.4.4
pdfkit==0.6.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
pipenv==2018.11.26
pkginfo==1.5.0.1
prompt-toolkit==3.0.0
protobuf==3.0.0
psutil==5.4.2
ptyprocess==0.6.0
pycairo==1.16.2
pycodestyle==2.5.0
pycparser==2.19
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.1
Pygments==2.5.1
pygobject==3.26.1
PyICU==1.9.8
pyinotify==0.9.6
pymacaroons==0.13.0
PyNaCl==1.1.2
Pyphen==0.9.5
pyRFC3339==1.0
python-apt==1.6.4
python-debian==0.1.32
python-magic==0.4.15
python-xapp==1.6.0
python-xlib==0.20
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
readme-renderer==24.0
reportlab==3.4.0
requests==2.22.0
requests-toolbelt==0.9.1
requests-unixsocket==0.1.5
restview==2.9.2
rss-reader==1.5.0
sessioninstaller==0.0.0
setproctitle==1.1.10
six==1.13.0
soupsieve==1.9.5
stevedore==1.31.0
system-service==0.3
systemd-python==234
tinycss==0.4
tinycss2==1.0.2
tqdm==4.38.0
traitlets==4.3.3
twine==2.0.0
ubuntu-drivers-common==0.0.0
ufw==0.36
urllib3==1.25.7
virtualenv==16.7.8
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
wcwidth==0.1.7
WeasyPrint==50
webencodings==0.5.1
xkit==0.0.0
yattag==1.12.2
